Feat: Create VenueArea controllers
Include functionality to create and remove Areas in a venue. Ability to create new storage space. Ability to set Sections, Shelves and Spots in a StorageSpace and update a Spot with new Product, par level or column span. Create controller for getting list of areas and getting a single area with full detail